Add files via upload
Hand gesture recognition is a computer vision project that involves the identification and interpretation of hand movements to perform various tasks. The project's goal is to develop a system that can recognize hand gestures in real-time and map them to corresponding actions.
The system will consist of a camera that captures the images of the user's hand gestures and a computer that will process these images using various computer vision algorithms to recognize and interpret the hand gestures. The system will use machine learning techniques to train a model that can accurately classify hand gestures.
This is created using OpenCV, Mediapipe, and tkinter libraries.
The tkinter library is used to create a graphical user interface (GUI) that will display the camera feed and processed image. The OpenCV library is used to capture video from the camera, and to process the image for hand gesture recognition. 
The Mediapipe library is used for detecting hand landmarks and finger tip positions.
The program starts by creating a GUI window with a frame for displaying the video feed.
 The to_pil function converts a frame from OpenCV format to PIL format, and displays it on a tkinter label. The FPS function is used to calculate the frames per second of the video feed. The run function captures a frame from the camera and resizes it to 600x500. It then uses the Mediapipe Hands model to detect the hand landmarks and finger tip positions. The paint function is used to draw lines between finger tip positions to show the path of the hand movement. The program continuously captures frames from the camera and processes them using the run function, which updates the video feed displayed in the GUI window. The program also displays the (x,y) coordinates of the index finger tip position in real-time.
Overall, this code can be used as a starting point for a hand gesture recognition project that involves tracking finger movements and recognizing gestures based on the finger positions. Overall, the hand gesture recognition project involves collecting data, pre-processing the data, extracting relevant features, training a machine learning model, and implementing real-time gesture recognition.